pyproject.toml
[tool.poetry]
name = "tdd-llm"
version = "0.1.0"
description = ""
authors = ["Muhammad Adnan Rizqullah <madnanrizqullah@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "^2.3.1"
pandas = "^2.3.0"
openai = "^1.93.0"
python-dotenv = "^1.1.1"
groq = "^0.30.0"
scipy = "^1.16.1"
pingouin = "^0.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
